+++ Plugins/Sentry/Source/Sentry/Private/GenericPlatform/GenericPlatformSentrySubsystem.cpp
@@ -309,16 +309,6 @@ sentry_value_t FGenericPlatformSentrySubsystem::OnBeforeMetric(sentry_value_t me
 
 sentry_value_t FGenericPlatformSentrySubsystem::OnCrash(const sentry_ucontext_t* uctx, sentry_value_t event, void* closure)
 {
-#ifdef USE_SENTRY_SESSION_REPLAY
-	if (SessionReplay && SessionReplay->HasSnapshotOnDisk())
-	{
-		TSharedPtr<ISentryAttachment> ReplayAttachment =
-			MakeShareable(new FGenericPlatformSentryAttachment(SessionReplay->GetAttachmentPath(), TEXT("session-replay.mp4"), TEXT("video/mp4")));
-
-		AddFileAttachment(ReplayAttachment);
-	}
-#endif
-
 	if (isScreenshotAttachmentEnabled && !IsOutOfProcessScreenshotEnabled() && !IsRunningCommandlet())
 	{
 		if (IsScreenshotSupported())
@@ -467,6 +457,11 @@ void FGenericPlatformSentrySubsystem::SetEventTag(sentry_value_t event, const ch
 		sentry_value_set_by_key(event, "tags", eventTags);
 	}
 	sentry_value_set_by_key(eventTags, key, sentry_value_new_string(value));
+}
+
+void FGenericPlatformSentrySubsystem::AttachStackTrace(sentry_value_t target)
+{
+	sentry_value_set_stacktrace(target, nullptr, 0);
 }
 
 FGenericPlatformSentrySubsystem::FGenericPlatformSentrySubsystem()
@@ -827,7 +822,7 @@ TSharedPtr<ISentryId> FGenericPlatformSentrySubsystem::CaptureMessage(const FStr
 
 	if (isStackTraceEnabled)
 	{
-		sentry_value_set_stacktrace(nativeEvent, nullptr, 0);
+		AttachStackTrace(nativeEvent);
 	}
 
 	sentry_uuid_t id = sentry_capture_event(nativeEvent);
@@ -840,7 +835,7 @@ TSharedPtr<ISentryId> FGenericPlatformSentrySubsystem::CaptureMessageWithScope(c
 
 	if (isStackTraceEnabled)
 	{
-		sentry_value_set_stacktrace(nativeEvent, nullptr, 0);
+		AttachStackTrace(nativeEvent);
 	}
 
 	sentry_scope_t* scope = sentry_local_scope_new();
@@ -862,7 +857,7 @@ TSharedPtr<ISentryId> FGenericPlatformSentrySubsystem::CaptureEvent(TSharedPtr<I
 
 	if (isStackTraceEnabled)
 	{
-		sentry_value_set_stacktrace(nativeEvent, nullptr, 0);
+		AttachStackTrace(nativeEvent);
 	}
 
 	sentry_uuid_t id = sentry_capture_event(nativeEvent);
@@ -877,7 +872,7 @@ TSharedPtr<ISentryId> FGenericPlatformSentrySubsystem::CaptureEventWithScope(TSh
 
 	if (isStackTraceEnabled)
 	{
-		sentry_value_set_stacktrace(nativeEvent, nullptr, 0);
+		AttachStackTrace(nativeEvent);
 	}
 
 	sentry_scope_t* scope = sentry_local_scope_new();
@@ -900,7 +895,7 @@ TSharedPtr<ISentryId> FGenericPlatformSentrySubsystem::CaptureEnsure(const FStri
 	sentry_value_t nativeException = sentry_value_new_exception(TCHAR_TO_UTF8(*type), TCHAR_TO_UTF8(*message));
 	sentry_event_add_exception(exceptionEvent, nativeException);
 
-	sentry_value_set_stacktrace(exceptionEvent, nullptr, 0);
+	AttachStackTrace(exceptionEvent);
 
 	FString ScreenshotPath;